move types folder to another location, change import to from `types`
