+++ types/interfaces/global/index.ts
